+++ global.d.ts
@@ -1,3 +1,4 @@
+import type {} from 'hexo'
 declare module '*.yml' {
     const data: any;
     export default data;