+++ src/helpers/my.ts
@@ -5,39 +5,47 @@ function gettingArr(parArray: any) {
     return undefined
 }
 
-hexo.extend.helper.register('my_post_main', function (point, post) {
+hexo.extend.helper.register('my_post_main', (point, post) => {
     const categories = gettingArr(post.categories)
     const tags = gettingArr(post.tags)
 
-    const res = hexo.execFilterSync(point, { categories, tags })
+    const res = hexo.execFilterSync(point, { categories, tags }, null)
     if (typeof res === 'string') {
         return res
     }
-    return null
+    return ''
 })
 
-hexo.extend.helper.register('my_post_js', function (post) {
+hexo.extend.helper.register('my_post_js', (post) => {
     const categories = gettingArr(post.categories)
     const tags = gettingArr(post.tags)
 
     const { source } = post
 
-    const res = hexo.execFilterSync('postJS', { categories, tags, source })
+    const res = hexo.execFilterSync(
+        'postJS',
+        { categories, tags, source },
+        null
+    )
     if (typeof res === 'string') {
         return res
     }
-    return null
+    return ''
 })
 
-hexo.extend.helper.register('my_post_css', function (post) {
+hexo.extend.helper.register('my_post_css', (post) => {
     const categories = gettingArr(post.categories)
     const tags = gettingArr(post.tags)
 
     const { source } = post
 
-    const res = hexo.execFilterSync('postCSS', { categories, tags, source })
+    const res = hexo.execFilterSync(
+        'postCSS',
+        { categories, tags, source },
+        null
+    )
     if (typeof res === 'string') {
         return res
     }
-    return null
+    return ''
 })