+++ src/utils/__test__/htmlEl.test.ts
@@ -1,0 +1,9 @@
+import { imgElGlobalClass } from '../htmlEl'
+
+describe('imgElGlobalClass', () => {
+    test('class', () => {
+        expect(imgElGlobalClass('<img alt="xxx" >')).toEqual(
+            '<img class="img-fluid object-fit-cover" alt="xxx" >'
+        )
+    })
+})