+++ src/utils/htmlEl.ts
@@ -1,0 +1,3 @@
+export function imgElGlobalClass(imgStr: string) {
+    return imgStr.replace('<img', '<img class="img-fluid object-fit-cover"')
+}