+++ src/tags/group-image.ts
@@ -9,7 +9,7 @@ const groupImage = (args: string[], content: string) => {
     let images: string[] = content.match(/<img[\s\S]*?>/g)!
     const rowCol = igRowCol(images.length)
     if (rowCol === 1) {
-        return imgElGlobalClass(images[0], ['img-swipe-single'])
+        return imgElGlobalClass(images[0], ['w-100 img-swipe-single'])
     }
     if (imgsSameSize) {
         images = images.map((item) => {