+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - 读取 hexo basedir / src 目录下的配置，分别写入到项目中
 - 启用 CSR 客户端渲染文章，开启后将去同步扫描文章对应的js目录，并读取目录下面的yml配置信息，读取到了，载入js和css，没有配置文件，则直接载入对应目录的js，默认都是esm，其他cdn的都当作是umd
 - index.d.ts LNote for browser
+- lnote-lightbox-caption 支持 img alt
 
 ### Changed
 
@@ -13,6 +14,7 @@
 - 关闭网页访问次数功能
 - 调整 nonu 模板，vue3已经支持适配html attr，如果绑定的值不存在，则不会拥有该 attr
 - 优化css，移除没用到的资源
+- js img-swipe use native size
 - ~~模板数据解析支持扫描目录下的所有数据，暂不支持，每份数据规则不同，需要考虑自行处理~~
 
 ### Fixed