+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+## 1.2.6
+
+### Added
+- local build and push scripts to git
+- git release and tigger git action to publish npm the new version
+
+### Fixed
+
+- img-swipe alt null not show title
+
+### Change
+- uri utils param sync
+
 ## 1.2.5
 
 ### Added
@@ -68,3 +81,6 @@
 ### Fixed
 
 ### Changed
+
+
+[git npm publish key](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)